get_sdk_release_change_log/requirements.txt
openai>=1.50.0
requests>=2.31.0
tenacity>=8.2.3
python-dateutil>=2.8.2
pyyaml>=6.0.1

